move and update commands/menu/event api
